Chore: Update Rails to 7.1 (#11)
* Upgrade to Rails 7.1 and update related dependencies

Upgraded Rails from 7.0 to 7.1, updated related gems (puma, rspec-rails) for compatibility, and adjusted application configuration and model serialization for Rails 7.1 changes. The ActiveRecord PostgreSQL decoder monkey patch is now skipped for Rails 7.0+ as it is no longer needed.

* Update Rails comment to specify version 7.1.6

Adjusted the Rails gem comment in the Gemfile to indicate upgrade to Rails 7.1.6 and clarified comments for Psych and Zeitwerk regarding Rails 7.0+ compatibility.

* Add .env.prod to .gitignore

Added .env.prod to the .gitignore file to prevent production environment variables from being committed to the repository.